multiple things:
trying to fix the date picker
trying to fix grunt tasks
started to make the image management section, updated the fetch script to input more keys from
the AL db, and now making the form to be able to complete this data when uploading directly.

Currently stuck on passing data from the add image form to the db.