Update method for image side data
Not entirely finished yet, the update image route needs to be able to show
the image we're updating, but at the moment just expects a new image to
upload. This won't do of course, so there may need to be a separate form
for this, or a condition, or a way to bypass the DataRequired value on the
update condition.

Also the title_and_year key was all messed up and included HTML and lots
of mess, but seemed to be a concatination of two other keys so I'm reverting
to those to ensure I can reproduce exactly the same datatype when adding new
image as to match image that will be imported from AL