The login form with WTF and flash classes
Built in the login form to integrate with WTForms to benefit from CSRF and bootstrap styles
and ease of mind and general integration with the rest of the admin.

Further, Flask flashes has an ability to register flash classes, and bootstrap comes built
in with a set of alet types, so it's an easy combo. We were not harnessing this up to now
though, so each flash message can now be appended with an extra argument, corresponding to
bootstrap alert classes:
http://getbootstrap.com/components/#alerts

Now loooking to put in an if statement is in place, in case we forget new arg, to revert
to 'alert-info'.