Finalizing the image meta data manipulation section.
We now have all the methods needed to add new, manipulate and delete
images and their associated data. Note that the update form is
slightly modified because I only wanted the user to be able to
change the meta data, I'm considering that the image itself is to
remain constant. If it is not, then it is to be deleted.